pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-interpreter-enterprise"
version = "0.1.0"
description = "Enterprise-ready data science starter repository for PDF interpretation and document intelligence."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "PDF Interpreter Team" }
]
dependencies = [
  "joblib>=1.4,<2.0",
  "pandas>=2.2,<3.0",
  "pyyaml>=6.0,<7.0",
  "scikit-learn>=1.5,<2.0",
  "typer>=0.12,<1.0"
]

[project.optional-dependencies]
dev = [
  "mypy>=1.11,<2.0",
  "pytest>=8.3,<9.0",
  "ruff>=0.6.9,<1.0"
]

[project.scripts]
pdf-interpreter = "enterprise_ds.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
packages = ["enterprise_ds"]
ignore_missing_imports = true
